pyproject.toml
[project]
name = "power-bi-flask-web-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-environ>=0.13.0",
    "flask>=3.1.3",
    "flask-smorest>=0.46.2",
    "gunicorn>=25.1.0",
    "marshmallow>=4.2.2",
]
